licensecheck: remove copyright lines from consideration
For many instances of license files, the only significant difference
between the reference and actual text is one or more copyrights,
almost always provided as a single line:

	Copyright Juan de Fuca 2020

This causes a number of problems, most notably poor confidence
scores when there are many copyright lines attached to a short
license, and also irrelevant variance in confidence scores depending
on the volume of copyright text.

Our solution is simple: Delete all copyright lines from the text
we process. But we cannot just delete them, as the Match type
includes location information, so we replace the text with spaces.