licensecheck: add Checker for defining custom license sets
Instead of forcing the use of a hard-coded builtin set, this lets clients of the
library define their own license sets. There is still a builtin set, but being able
to define alternate sets should reduce the pressure to add every last possible
license to the built-in set and let users adjust the set without forking the
module.

Fixes #10.