licensecheck: expand match Start/End to text start/end when appropriate
If a match starts at word 0 or ends at the last word, it makes sense
to extend the text byte offsets to the start or end of the text, respectively.
There is already a special case in the URL matcher to do this at the end
of the text, but there wasn't one for the main license text matches.

Adding this case makes the returned match offsets cover the whole
file for a whole-file match, as seen in the testdata changes.

Removed TestMatch because it is redundant with the now many test
cases in the testdata directory and was fussy about the offsets changing.