licensecheck: fix backward gap check going too far
Looking backward across a gap for the word Copyright is too easy
to jump into the previous license. So stop at the first one.

There may be a better way to do this by running the license matches
in parallel instead of series, which would let us know when we were
scanning backward into a previous license, but for now this should be OK.

Fixes #15.