internal/match: add dictionary implementation for license scans
This dictionary is a cleaned up version of the one in licensecheck proper,
and it adds new methods InsertSplit and Split, which are analogous to
normalize. Unlike normalize, which returns a list of word IDs ([]int),
InsertSplit and Split returns list of Word structures with precise location
information. Also unlike normalize, InsertSplit and Split convert words
to lowercase one at a time in a temporary buffer, allowing proper
implementation of Unicode lowercasing even in the presence of rune size
changes.

Conversion runs at about 50 MB/s.