licenses: add GooglePatentClause, GooglePatentsFile, and CommonsClause [LRE]
GooglePatentClause is the original patent text added to the Go LICENSE file,
still present in early forks of the code.

GooglePatentsFile is the Go PATENTS file that replaced that clause,
also used in other BSD-licensed Google projects like WebM, gRPC, and so on.
It has also been adapted to other companies in a few instances.

CommonsClause is the Commons Clause, which we want to recognize
and report, since it is problematic for many uses.