internal/match: add implicit cut operator to reduce scope of wildcards
Quoting the new comment in rematch.go:

Early-Cut Wildcard Matching

This implementation adds "cut" operations to reduce the number
of states tracked when using counted wildcard patterns.

To explain what that means exactly, first some background.

DFAs are not good at counting: the standard regexp pattern /.{0,10}/
is much more expensive than /.*/. On the other hand, /.*/ can easily
match far too much; the limit in /.{0,10}/ is semantically useful.

LREs provide a counted wildcard __10__ equivalent to regexp /.{0,10}/.
The counting can produce a multiplicative number of DFA states.
For example, consider the pattern "The name __10__ may not be used"
matched against "The name may not be may not be may not be may not be used".
If __10__ means "up to 10 words, any words at all", then that text should
be matched, assigning "may not be may not be may not be" to the wildcard.
This never happens in practice, but the standard DFA construction must be
prepared for it, along with variants like:

	The name may may may may may may may may may may may not be used.
	The name may not may not may not may not may not may not be used.
	The name may may not be  may not be  may not be  may not be used.
	The name may not may not be used may not be used may not be used.

In general the DFA must track
(1) how many wildcard words have been used up so far, and
(2) how far into the "may not be used" has been matched
at the current input location. That's roughly 10 * 4 = 40 states.
In a real license, the possible literal text that must be tracked
is as many words as the wildcard itself allows, leading in this case
to roughly 10 * 10 = 100 states. A __20__ produces roughly 400 states,
and so on.

This extra work is clearly useless: real inputs don't look like that,
and people wouldn't recognize them if they did. After seeing a few of
the words following the wildcard, we can safely assume that the wildcard
part of the match is over. In the example above, after seeing

	The name Google may not be

we might as well assume the wildcard is over and the next word must
be "used", to complete the pattern. This is true even though technically
the pattern might be interpreted to allow

	The name "Google may not be may not be" may not be used.

Empirically, this kind of abuse does not come up in practice.

This implementation cuts off wildcard matches by inserting an implicit
"cut" operator (similar to and named after Prolog's cut operator)
three literal words after each wildcard. The cut discards (cuts off)
any NFA threads still attempting wildcard matches.
That is, our example is interpreted implicitly as

	The name __10__ may not be (CUT) used.

The effect of this cut after three words is that the 10 * 10 states
drops to 10 * 3, 20 * 20 drops to 20 * 3, and so on: wildcards now
have a footprint only linear in their size, not quadratic.

As a larger example, the implicit cut in the longer pattern

	The name __20__ may not be (CUT) used to endorse or promote products
	derived from this software without specific prior written permission.

reduces the total number of DFA states from 248 to 80.

It is still possible to delay the cut by following the wildcard with
optional words or phrases, so some state blowup is still possible,
but not nearly as much.

Overall, at time of writing, implicit cuts reduce the size of the
DFA for the full license set from 5.8M states (240 MB and 39s to build)
to 615k states (10 MB, under one second to build).

The implicit cut is represented as an instCut that terminates any
other NFA threads still matching a particular wildcard. The application
of instCut happens in (*nfaState).trim.