licensecheck: add Trace function
Add the Trace function, which sets the match.TraceDFA variable.

This is invaluable for debugging uses of this package.

As one example, you are faced with what seems to be a fine Apache-2.0
license that nonetheless isn't matched by any LRE. You try to compare
the text to the LRE, but quickly realize that this task is toilsome
and inaccurate. Calling Trace and running again immediately reveals
that someone has s/product/bundle/, resulting in "reproduction" being
altered to "rebundleion". (True story.)

Trace is also useful when writing custom LREs to capture new license
variants.